mcp-monitor/app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
requests==2.32.3
pydantic==2.9.2
PyYAML==6.0.2
